feat: restructure details generator
