--- nx-plugin/src/generators/search/steps/default.step.ts
@@ -1,12 +1,0 @@
-import { Tree } from '@nx/devkit';
-import { GeneratorStep } from '../../shared/generator.utils';
-import { SearchGeneratorSchema } from '../schema';
-
-export class DefaultStep implements GeneratorStep<SearchGeneratorSchema> {
-  process(tree: Tree, options: SearchGeneratorSchema): void {
-    throw new Error('Method not implemented.');
-  }
-  getName(): string {
-    throw new Error('Method not implemented.');
-  }
-}
